Completed Traits and Main
